pyproject.toml
[tool.poetry]
name = "favihunter"
version = "1.1.0"
description = "Discover and track internet assets using favicon hashes through search engines."
authors = ["eremit4 <eremit4@protonmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/eremit4/favihunter"
homepage = "https://github.com/eremit4/favihunter"

[tool.poetry.dependencies]
python = "^3.8"
mmh3 = "4.0.1"
favicon = "0.7.0"
colorama = "0.4.6"
requests = "2.32.3"
fake-useragent = "1.5.1"
pillow = "10.4.0"
PyYAML = "6.0.2"

[tool.poetry.scripts]
favihunter = "favihunter.favihunter:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"